agent.py
"""
Main script for the Stock Analyst Agent.

This script sets up a multi-agent system for financial analysis, including a supervisor agent and specialized agents for market data, financial analysis, and news.
"""

import os
from dotenv import load_dotenv
from typing import Literal
from langchain_groq import ChatGroq
from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
from pydantic import BaseModel

# Load environment variables from .env file
load_dotenv()

# Load environment variables
GROQ_API_KEY = os.getenv("GROQ_API_KEY")
TAVILY_API_KEY = os.getenv("TAVILY_API_KEY")
GROQ_MODEL = os.getenv("GROQ_MODEL")

# LLM definition
llm = ChatGroq(model=GROQ_MODEL, api_key=GROQ_API_KEY)

# Define the structure for the supervisor's decision
class SupervisorDecision(BaseModel):
    """
    Represents the decision of the supervisor agent, indicating the next agent to act or to finish the process.
    """
    next: Literal["Market_Data_Agent", "Analysis_Agent", "News_Agent", "FINISH"]

# Define agent members
AGENT_MEMBERS = ["Market_Data_Agent", "Analysis_Agent", "News_Agent"]

# Supervisor prompt setup
SUPERVISOR_SYSTEM_PROMPT = (
    "You are a Financial Service Supervisor managing the following agents: "
    f"{', '.join(AGENT_MEMBERS)}. Select the next agent to handle the current query."
)

SUPERVISOR_PROMPT = ChatPromptTemplate.from_messages([
    ("system", SUPERVISOR_SYSTEM_PROMPT),
    MessagesPlaceholder(variable_name="messages"),
    ("system", "Choose the next agent from: {options}."),
]).partial(options=str(AGENT_MEMBERS))

# Supervisor Agent
def supervisor_agent(state):
    """
    The supervisor agent that routes the query to the appropriate agent.
    """
    supervisor_chain = SUPERVISOR_PROMPT | llm.with_structured_output(SupervisorDecision)
    return supervisor_chain.invoke(state)

# --- Missing Agent and Graph Definitions ---
# The following sections are placeholders for the actual agent and graph definitions,
# which are not present in the provided code.

# (Placeholder for Market_Data_Agent definition)

# (Placeholder for Analysis_Agent definition)

# (Placeholder for News_Agent definition)

# (Placeholder for StateGraph definition and compilation)
