requirements.txt
python-dotenv
yfinance
langchain
langchain-community
langchain-groq
langchain-core
pydantic
langgraph
groq
tavily-python
